apps/api/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
gunicorn==21.2.0
supabase==2.8.0
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1

